Part4: Exercises 4.15 to 4.23 - Bloglist (Authorization with JWT)

+++ part4/blog_list/controllers/blogs.js
@@ -1,10 +1,10 @@
 const router = require('express').Router()
 const Blog = require("../models/blog")
 
-router.get('/', (request, response) => {
-    Blog.find({}).then(blogs => {
-         response.status(200).json(blogs)
-    })
+router.get('/', async(request, response) => {
+    const blogs = await Blog.find({}).populate("author", {name: 1})
+
+    response.status(200).json(blogs)
 })
 
 router.get("/:id", (request, response, next) => {
@@ -14,25 +14,32 @@ router.get("/:id", (request, response, next) => {
     .catch(error => next(error))
 })
   
-router.post('/', (request, response, next) => {
-    const blogBody = request.body
+router.post('/', async(request, response, next) => {
+    let blogBody = request.body
+    const localUser = request.user
 
     if(!blogBody.title || !blogBody.url) return response.status(400).json({ error: "Title or URL are missing" })
-
-    if(!blogBody.likes) blogBody.likes = 0
+    if(!blogBody.likes) blogBody = {...blogBody, likes: 0}
+    blogBody = {...blogBody, author: localUser.id}
 
     const blog = new Blog(blogBody)
   
-    blog.save().then(result => {
+    try{
+        const result = await blog.save()
         response.status(201).json(result)
-    })
-    .catch(error => next(error))
+    } catch(e){next(e)}
 })
 
 router.delete("/:id", async(request, response) => {
-    const blogDeleted = await Blog.findByIdAndDelete(request.params.id)
+    const blogtoDelete = await Blog.findById(request.params.id)
+    const localUser = request.user
+    if(localUser.id === blogtoDelete.author.toString()){
+        const blogDeleted = await Blog.findByIdAndDelete(request.params.id)
     
-    return response.status(204).json(blogDeleted)
+        return response.status(204).json(blogDeleted)
+    } else {
+        return response.status(403).json({ error: "Permission denied" })
+    }
 })
 
 router.put("/:id", async(request, response) => {
